Moved main_menu to it's own folder
