+++ src/main_menu/resources.rs
@@ -1,0 +1,16 @@
+use bevy::prelude::*;
+
+#[derive(Resource)]
+pub struct ButtonColors {
+    pub normal: Color,
+    pub hovered: Color,
+}
+
+impl Default for ButtonColors {
+    fn default() -> Self {
+        ButtonColors {
+            normal: Color::rgb(0.15, 0.15, 0.15),
+            hovered: Color::rgb(0.25, 0.25, 0.25),
+        }
+    }
+}
